requirements.txt
fastmcp>=2.11.0
fastapi
uvicorn
python-dotenv
aiohttp
